Moved draw functionality out of MainController and into new class SpriteMainMenuView.java.
